==> Datasets/Independent Datasets - PanEtAl/H.Pylori/1. Creating FASTA file for CD-Hit.ipynb <==
import pandas as pd
import numpy as np
import math
import json
import os
import scipy.io, random
from pathlib import Path
import itertools

abs_path = os.getcwd()

data_path = str(Path().resolve().parent) + "/PPI_sequence/"

data_path

PAmat = scipy.io.loadmat(data_path + 'Mix_Hpylo_data.mat')
PBmat = scipy.io.loadmat(data_path + 'Mix_Hpylo_data.mat')

P1 = [PAmat['protein_A'][i, 0][0] for i in range(len(PAmat['protein_A']))]
P2 = [PBmat['protein_B'][i, 0][0] for i in range(len(PBmat['protein_B']))]

unique = list(dict.fromkeys(list(itertools.chain(P1, P2))))

protdict = {}
for i in range(len(unique)):
    protdict[unique[i]] = "P" + str(i)

PA = [protdict[prot] for prot in P1]
PB = [protdict[prot] for prot in P2]

#rand_nums = random.sample(list(range(len(NAmat['N_protein_a']))), len(PAmat['P_protein_a']))
#N1 = [NAmat['N_protein_a'][i, 0][0] for i in rand_nums]
#N2 = [NBmat['N_protein_b'][i, 0][0] for i in rand_nums]

pdict = [{'pA': pA, 'pB': pB} for pA, pB in zip(PA, PB)]

protdict_map = {v: k for k, v in protdict.items()}

with open("positive.json", 'w') as outfile:
    json.dump(pdict, outfile)   

df_pos = pd.read_pickle(str(Path().resolve().parent.parent) + "/Benchmark/PanEtAl/positive.pkl")
df_neg = pd.read_pickle(str(Path().resolve().parent.parent) + "/Benchmark/PanEtAl/negative.pkl")

total_proteins_bench = set(df_neg['Protein_1']) | set(df_pos['Protein_1']) | set(df_neg['Protein_2']
                                                                                ) | set(df_pos['Protein_2'])

total_proteins_bench = list(total_proteins_bench)
len(total_proteins_bench)

new_df = df_neg[['Protein_1', 'Protein_seq1']]
new_df.columns = ['ID', 'Sequence']
temp = df_neg[['Protein_2', 'Protein_seq2']]
temp.columns = ['ID', 'Sequence']
new_df = new_df.append(temp, ignore_index = True)
temp = df_pos[['Protein_1', 'Protein_seq1']]
temp.columns = ['ID', 'Sequence']
new_df = new_df.append(temp, ignore_index = True)
temp = df_pos[['Protein_2', 'Protein_seq2']]
temp.columns = ['ID', 'Sequence']
new_df = new_df.append(temp, ignore_index = True)

benchdict = new_df.set_index('ID')['Sequence'].to_dict()

res = {**benchdict, **protdict_map}

ofile = open("bench_hpylori.fasta", "w")

for protein in res.keys():
    ofile.write(">" + protein + "\n" + res[protein] + "\n")

ofile.close()

with open("seq.json", 'w') as outfile:
    json.dump(res, outfile)

# For 40%
http://weizhong-lab.ucsd.edu/cdhit_suite/cgi-bin/result.cgi?JOBID=1654178430
