==> Datasets/Independent Datasets - PanEtAl/H.Pylori/2. Parsing CD-Hit files.ipynb <==
import re, json
import pandas as pd
import pickle
import networkx as nx

file1 = open('1654178430.result/1654178430.fas.1.clstr.sorted', 'r')
Lines = file1.readlines()

LineIndex = 0
cd_hit_dict = {}
while LineIndex < len(Lines):
    FF = re.search(r'^(>Cluster.*)', Lines[LineIndex])
    if FF != None:
        LineIndex += 1
        li = []
        dkey = ''
        while LineIndex < len(Lines):
            KK = re.search(r'^(.*\*)', Lines[LineIndex])
            if KK != None:
                dkey = KK.group()
                dkey = re.sub(r'.*>','',dkey)
                dkey = re.sub(r'\.\.\. \*','',dkey)
            else:
                KK = re.search(r'^(.*\.\.\.)', Lines[LineIndex])
                dval = KK.group()
                dval = re.sub(r'.*>','',dval)
                dval = re.sub(r'\.\.\.','',dval)
                li.append(dval)
            if(LineIndex == len(Lines)-1):
                break
            FF = re.search(r'^(>Cluster.*)', Lines[LineIndex + 1])
            if (FF == None):
                LineIndex += 1
            else:
                break
        cd_hit_dict.update({dkey: li})
    LineIndex += 1

li = []
for k in cd_hit_dict.keys():
    if(k.startswith('P')):
        if(not(cd_hit_dict[k])):
            li.append(k)
        elif(all(protein.startswith('P') for protein in cd_hit_dict[k])):
            li.append(k)
            li.extend(cd_hit_dict[k])
len(li)

f = open ('positive.json', "r")
 
p_data = json.load(f)

f.close()

pA = []
pB = []
pgr = nx.Graph()
for i in range(len(p_data)):
    pA.append(p_data[i]['pA'])
    pB.append(p_data[i]['pB'])
    pgr.add_edge(p_data[i]['pA'], p_data[i]['pB'])

to_remove = list(set(list(pgr.nodes())) - set(li))

for i in range(len(to_remove)):
    pgr.remove_node(to_remove[i])

f = open ('seq.json', "r")
 
seq_dict = json.load(f)

f.close()

seq_selected = {x : seq_dict[x] for x in li}

nx.to_pandas_edgelist(pgr)

with open('hpylori.pkl', 'wb') as f:
    pickle.dump([nx.to_pandas_edgelist(pgr), seq_dict], f)
